Renamed ngraph_function creation samples
