+++ samples/cpp/model_creation_sample/CMakeLists.txt
@@ -2,9 +2,9 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-set(TARGET_NAME "ngraph_function_creation_sample")
+set(TARGET_NAME "model_creation_sample")
 
-ie_add_sample(NAME ngraph_function_creation_sample
+ie_add_sample(NAME model_creation_sample
               SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/main.cpp"
-              HEADERS "${CMAKE_CURRENT_SOURCE_DIR}/ngraph_function_creation_sample.hpp"
+              HEADERS "${CMAKE_CURRENT_SOURCE_DIR}/model_creation_sample.hpp"
               DEPENDENCIES format_reader ie_samples_utils)
