+++ samples/cpp/model_creation_sample/lenet.labels
