+++ samples/cpp/model_creation_sample/model_creation_sample.hpp
