+++ samples/python/model_creation_sample/data.py
