+++ samples/python/model_creation_sample/lenet.bin
