+++ samples/python/model_creation_sample/model_creation_sample.py
